rpc/requirements.txt
yadisk_async==1.3.3
Pillow==9.2.0
aio-pika==8.1.0
pydantic==1.9.1
python-dotenv==0.20.0


